moved json calibration methods to helpers.py
